Fixed an issue with batch requests being faster than unary requests. Added the use of serilog for more effective debugging, getting ready for updating the database, also cleaned up the code a little bit more.

+++ SharedCommonalities/ReturnModels/ReturnTypes/ClientDetails.cs
@@ -11,7 +11,7 @@ namespace SharedCommonalities.ReturnModels.ReturnTypes
 
         public Guid RequestId { get; set; }
 
-        public int MessageLength { get; set; }
+        public long MessageLength { get; set; }
 
         public bool IsActiveClient { get; set; }
